move component to folder components/EventCard
