+++ src/components/EventCard/EventCard.css
